compilers: move CFLAGS/CXXFLAGS handling to Environment
That is where env_opts are stored, so make the compiler call back directly
into the environment.

Suggested-by: Dylan Baker <dylan@pnwbakers.com>
